- (snmp_api.h, snmp_api.c, snmp_auth.c, snmp_impl.h, asn1.c): 64bit cleanup.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@3049 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_impl.h
@@ -148,8 +148,8 @@ extern int snmp_errno;
  */
 #define FIRST_PASS	1
 #define	LAST_PASS	2
-u_char	*snmp_comstr_parse(u_char *, size_t *, u_char *, size_t *, int *);
-u_char	*snmp_comstr_build (u_char *, size_t *, u_char *, size_t *, int *, size_t);
+u_char	*snmp_comstr_parse(u_char *, size_t *, u_char *, size_t *, long *);
+u_char	*snmp_comstr_build (u_char *, size_t *, u_char *, size_t *, long *, size_t);
 
 int has_access (u_char, int, int, int);
 #ifdef __cplusplus
